commit
розділив функції на окремі файли
