+++ src/index.js
@@ -5,3 +5,4 @@ import './js/weather';
 import './js/listener';
 import './js/search';
 import './js/dark_mode';
+import './js/readMore';
